CASSANDRA-17964 4.0
