+++ test/unit/org/apache/cassandra/cql3/KeywordSplit2Test.java
@@ -28,9 +28,9 @@ import org.junit.runners.Parameterized;
  * KeywordTestSplitN to prevent CI timing out. If timeouts reappear split it further
  */
 @RunWith(Parameterized.class)
-public class KeywordTestSplit1 extends KeywordTestBase
+public class KeywordSplit2Test extends KeywordTestBase
 {
-    static int SPLIT = 1;
+    static int SPLIT = 2;
     static int TOTAL_SPLITS = 2;
     
     @Parameterized.Parameters(name = "keyword {0} isReserved {1}")
@@ -38,7 +38,7 @@ public class KeywordTestSplit1 extends KeywordTestBase
         return KeywordTestBase.getKeywordsForSplit(SPLIT, TOTAL_SPLITS);
     }
     
-    public KeywordTestSplit1(String keyword, boolean isReserved)
+    public KeywordSplit2Test(String keyword, boolean isReserved)
     {
         super(keyword, isReserved);
     }
